refactor(memo): simplify memo tool argument/scope helpers
